+++ python/AIEMLIRModule.cpp
@@ -15,7 +15,7 @@ using namespace mlir::python::adaptors;
 
 PYBIND11_MODULE(_aie, m) {
 
-  ::aieRegisterAllPasses();
+  aieRegisterAllPasses();
 
   m.doc() = R"pbdoc(
     AIE MLIR Python bindings
@@ -43,7 +43,7 @@ PYBIND11_MODULE(_aie, m) {
   mlir_type_subclass(m, "ObjectFifoType", aieTypeIsObjectFifoType)
       .def_classmethod(
           "get",
-          [](py::object cls, MlirType type) {
+          [](const py::object &cls, const MlirType type) {
             return cls(aieObjectFifoTypeGet(type));
           },
           "Get an instance of ObjectFifoType with given element type.",
@@ -52,7 +52,7 @@ PYBIND11_MODULE(_aie, m) {
   mlir_type_subclass(m, "ObjectFifoSubviewType", aieTypeIsObjectFifoSubviewType)
       .def_classmethod(
           "get",
-          [](py::object cls, MlirType type) {
+          [](const py::object &cls, const MlirType type) {
             return cls(aieObjectFifoSubviewTypeGet(type));
           },
           "Get an instance of ObjectFifoSubviewType with given element type.",
